feat: create special dtos for seat controller
