+++ src/main/java/org/vr61v/mappers/SeatMapper.java
@@ -2,14 +2,17 @@ package org.vr61v.mappers;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import org.vr61v.dtos.SeatDto;
+import org.vr61v.dtos.seat.SeatRequestDto;
+import org.vr61v.dtos.seat.SeatResponseDto;
 import org.vr61v.entities.Seat;
 
 @Mapper(componentModel = "spring")
-public interface SeatMapper extends BaseMapper<Seat, SeatDto> {
+public interface SeatMapper {
 
     @Mapping(target = "aircraftCode", source = "entity.id.aircraft.aircraftCode")
     @Mapping(target = "seatNo", source = "entity.id.seatNo")
-    SeatDto toDto(Seat entity);
+    SeatResponseDto toDto(Seat entity);
+
+    Seat toEntity(SeatRequestDto dto);
 
 }
